Fixing with a temporary fix the go.uuid <> sqlx runner issues
Recently [1], satori/go.uuid changed the signature of various functions, including NewV4 used by sqlx runner. As a result, sqlx runner does not compile at HEAD anymore. A change was submitted to sqlx runner [2], but meanwhile, to keep our build green, we're forcibly setting the satory/go.uuid version to a specific sha. This approach is chosen over vendoring since the intent of worksheets is to be used as a library in a project which would itself manage vendoring.

[1] https://github.com/satori/go.uuid/commit/0ef6afb2f6cdd6cdaeee3885a95099c63f18fc8c
[2] https://github.com/mgutz/dat/pull/77