Being go-ish and requiring explicit return
This is a stepping stone to introducing more statements such as `if`, `for`, `x := expr`, etc. When we do introduce them, we can model them as optionally value producing expressions, and add a `ReturnsValue() bool` func to interrogate whether they do or do not, and require the top-level expression of a `computed_by` to produce a value.