A solution for sums needing to be initialized at 0
We had the issue where the `sum` in the following worksheet

	worksheet sum_example {
		1:nums []number[0]
		2:sum number[0] computed_by {
			sum(nums)
		}
	}

would be `undefined` when initiliazed, and when the first number is appended to the `nums`, it would actually represent the sum of numbers.

However, we've seen in practice that we want the sum to be defined to `0` when initializing the worksheet, as this actually represents the initial state.

To achieve this, suggesting two changes:

1. Slice fields start empty (vs undefined). Currently, only appending and deleting from slices is allowed, and we're therefore already saying that slice fields are never undefined, rather empty. This change makes it concrete in the internals.

2. When creating a new instance of a worksheet, we trigger all `computed_by` at least once. Most of these will yield `undefined` on `undefined` inputs, but some like sums would actually yield a result. We've already seen cases of running `computed_by` during initialization to support the "signoff pattern" (see `TestSignoffPattern`), so this is a generalization which fits in the spirit of the framework.