Adding min and max (and slice) (#118)
tl;dr see `rFoldNumbers` and then how this generic helper is used to write `min`, `max`, `sum`. Adding a `slice` func to create slices inline, which helps with testing.

Adding helper `slice` to create slices, and adding `min` and `max` which works over a wide range of numerical arguments: trying to allow any possible ways in which numbers could be passed, list of numbers, slices of numbers, slices of slices of numbers, a combination of all of the above. For that, introducing a way to 'fold' numerical arguments which is a just a way to iterate over all numerical arguments in a generic way.

Changing `sum` to be based over the same scaffolding which `min` and `max` are based on.