Support for computing averages, with avg (#120)
Adding new function `avg`, which introduces some additional complexity: since an average can be calculated to a specific precision, and in the spirit of worksheets where precision and rounding must be explicitly provided, we need to have the ability to
  1. Pass a rounding mode to a functional call (practically, attach a `tRound` to a `tCall`)
  2. Allow the function implementation to decide whether a rounding mode is required or not

Doing this in a few steps:
- In the parser, explicitly looking for a parsing mode following a function call, and directly attaching this to the `tCall` node
- In the `compute` of `tCall`, implementing rounding to preserve the functionality we had before
- Extending `fnArgs` to also have a rounding mode, such that individual functions can choose to do something with the rounding mode explicitely

While this works, this is pretty mediocre: rounding code is now in multiple places (in `compute` of `tBinop` and `tCall`), we round averages even though the rounding is already handled in the final division, some similarly wrong expressions have differing error messages. A more "pure" treatment seems like it'd require quite a bit of re-work, and re-shuffling, which doesn't feel worthwhile now. (And this re-work could likely handle the other complex rounding cases left un-implemented.)