Track References to Worksheets Through Time (#130)
This change introduces a more precise way to track references from parent worksheets to children worksheets. Before, we only stored the reference to the child, but lost the version of the child at the time this reference was recorded. Now, we additionally track the version of the child. This makes it possible to reconstruct precisely the worksheet graph, exactly as it was when recorded.

For instance, let's say we have an application worksheet, pointing to a customer worksheet. When we look at a prior version of an application, all we would know is that it pointed to customer with id XYZ. As such, a historical load of the application would have to guess which version of the customer to load...

To achieve this more precise recording, we first change how worksheet references are stored, by appending the version number to them. These now follow the pattern

    *:<uuid>@<version>

e.g `*:e310c9b6-fc48-4b29-8a66-eeafa9a8ec16@23`.

We modify the signature of `dbReadValue` to return both the value 'at head' (as it did before), as well as the value actually recorded (possibly lagging head). Having both of these values is important later when diffing, to properly record changes in the version of children worksheets.

To represent "a worksheet reference at version N", we introduce a special marker value `wsRefAtVersion`. This value isn't the worksheet at a specific version, it is _only_ representing the worksheet reference at a specific verrsion.

When diffing, we need to consider a `wsRefAtVersion` to be the same as the worksheet at the specific version, e.g. if we have the worksheet reference `e310c9b6-fc48-4b29-8a66-eeafa9a8ec16@23` is should be the same from a diffing standpoing as the worksheet `e310c9b6-fc48-4b29-8a66-eeafa9a8ec16` at version `23`.

To support this, we introduce `diffCompare` as a tailored comparison specifically for diffing. We leave `Equal` unchanged, as this is provided for semantic value equality.

For clarity, the code around diffing is moved into `diff.go` (and tests in `diff_test.go`).

Lastly, the overall feature testing lives in `refs_test.go`. Thes new tests are long, because they have assertions about the db state table by table, but have step-by-step explanations. Note that for backwards compatibility, we still want to support references without the version. There are specific tests ensuring this is handled in a backwards compatible way.